move qr code modal out of actions menu
